Refactor again to build with a separate service module
