+++ brotli4j/pom.xml
@@ -29,6 +29,14 @@
     <artifactId>brotli4j</artifactId>
     <packaging>jar</packaging>
 
+    <dependencies>
+        <dependency>
+            <groupId>com.aayushatharva.brotli4j</groupId>
+            <artifactId>service</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
+    </dependencies>
+
     <profiles>
         <profile>
             <id>linux-x86_64</id>
@@ -46,6 +54,7 @@
                 </dependency>
             </dependencies>
         </profile>
+
         <profile>
             <id>linux-aarch64</id>
             <activation>
@@ -58,6 +67,23 @@
                 <dependency>
                     <groupId>com.aayushatharva.brotli4j</groupId>
                     <artifactId>native-linux-aarch64</artifactId>
+                    <version>${project.parent.version}</version>
+                </dependency>
+            </dependencies>
+        </profile>
+
+        <profile>
+            <id>linux-armv7</id>
+            <activation>
+                <os>
+                    <family>Linux</family>
+                    <arch>arm</arch>
+                </os>
+            </activation>
+            <dependencies>
+                <dependency>
+                    <groupId>com.aayushatharva.brotli4j</groupId>
+                    <artifactId>native-linux-armv7</artifactId>
                     <version>${project.parent.version}</version>
                 </dependency>
             </dependencies>
@@ -137,11 +163,12 @@
                             <module>
                                 <moduleInfoSource>
                                     module ${javaModuleName} {
+                                        requires com.aayushatharva.brotli4j.service;
                                         exports com.aayushatharva.brotli4j;
                                         exports com.aayushatharva.brotli4j.common;
                                         exports com.aayushatharva.brotli4j.decoder;
                                         exports com.aayushatharva.brotli4j.encoder;
-                                        uses com.aayushatharva.brotli4j.BrotliNativeProvider;
+                                        uses com.aayushatharva.brotli4j.service.BrotliNativeProvider;
                                     }
                                 </moduleInfoSource>
                             </module>
