+++ brotli4j/src/main/java/com/aayushatharva/brotli4j/Brotli4jLoader.java
@@ -17,6 +17,7 @@
 package com.aayushatharva.brotli4j;
 
 import com.aayushatharva.brotli4j.common.annotations.Local;
+import com.aayushatharva.brotli4j.service.BrotliNativeProvider;
 
 import java.io.File;
 import java.io.InputStream;
