+++ brotli4j/src/test/java/com/aayushatharva/brotli4j/encoder/BrotliOutputStreamTest.java
