+++ brotli4j/src/test/resources/sample_data.txt
