+++ natives/pom.xml
@@ -39,7 +39,7 @@
     <dependencies>
         <dependency>
             <groupId>com.aayushatharva.brotli4j</groupId>
-            <artifactId>brotli4j</artifactId>
+            <artifactId>service</artifactId>
             <version>${project.parent.version}</version>
             <scope>provided</scope>
         </dependency>
